# SWKOM-paperless/wwwroot/openapi-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Mock Server",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/api/token" : {
      "post" : {
        "operationId" : "GetToken",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserInfo"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserInfo"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserInfo"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "tags" : [ "Login" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "UserInfo" : {
        "additionalProperties" : false,
        "example" : {
          "password" : "password",
          "username" : "username"
        },
        "properties" : {
          "username" : {
            "nullable" : true,
            "type" : "string"
          },
          "password" : {
            "nullable" : true,
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    }
  }
}